+++ db/schemas/net.bible.android.database.EpubDatabase/1.json
@@ -2,7 +2,7 @@
   "formatVersion": 1,
   "database": {
     "version": 1,
-    "identityHash": "2107ed889e5b741361279ad71050ca17",
+    "identityHash": "6a88dd1c437b4e03edf975f0e5edacf1",
     "entities": [
       {
         "tableName": "EpubHtmlToFrag",
@@ -32,11 +32,11 @@
       },
       {
         "tableName": "EpubFragment",
-        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`originalHtmlFileName` TEXT NOT NULL, `ordinalStart` INTEGER NOT NULL, `ordinalEnd` INTEGER NOT NULL, `id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL)",
+        "createSql": "CREATE TABLE IF NOT EXISTS `${TABLE_NAME}` (`originalId` TEXT NOT NULL, `ordinalStart` INTEGER NOT NULL, `ordinalEnd` INTEGER NOT NULL, `id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL)",
         "fields": [
           {
-            "fieldPath": "originalHtmlFileName",
-            "columnName": "originalHtmlFileName",
+            "fieldPath": "originalId",
+            "columnName": "originalId",
             "affinity": "TEXT",
             "notNull": true
           },
@@ -72,7 +72,7 @@
     "views": [],
     "setupQueries": [
       "CREATE TABLE IF NOT EXISTS room_master_table (id INTEGER PRIMARY KEY,identity_hash TEXT)",
-      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '2107ed889e5b741361279ad71050ca17')"
+      "INSERT OR REPLACE INTO room_master_table (id,identity_hash) VALUES(42, '6a88dd1c437b4e03edf975f0e5edacf1')"
     ]
   }
 }
